Change Container::all use return vector of descriptors
NWSync is a little problematic.. and punting on modified times for static containers for now

+++ lib/nw/resources/ResourceDescriptor.hpp
@@ -8,7 +8,7 @@ struct Container;
 
 struct ResourceDescriptor {
     Resource name;
-    int64_t size = 0;
+    size_t size = 0;
     int64_t mtime = 0;
     Container* parent = nullptr;
